Completely redesigned the DataFrame. Previously it would have a ton of image arrays listed to 1 row. This WRONG; changed it so that 1 row has exactly 1 name of butterfly and 1 array.
Tweaked some more lines in model.py, with these changes, we are able to have a working model with a tested 60% accuracy (terrible) with batch_size=64, epochs=12 and a training set of 20% of the population.

I changed the batch size to 128, and reduced the testing set to be 20% of the population and added the ability to save the model.